first version of refactor; sims still seem to run
